Logout buttons

+++ src/main/java/com/bestposts/blog/controllers/MainController.java
@@ -4,14 +4,14 @@ package com.bestposts.blog.controllers;
         import com.bestposts.blog.models.Posts;
         import com.bestposts.blog.models.Users;
         import com.bestposts.blog.repos.PostsRepo;
+        import com.bestposts.blog.repos.UsersRepo;
         import org.springframework.beans.factory.annotation.Autowired;
+        import org.springframework.security.core.userdetails.User;
         import org.springframework.stereotype.Controller;
         import org.springframework.ui.Model;
-        import org.springframework.web.bind.annotation.GetMapping;
-        import org.springframework.web.bind.annotation.PathVariable;
-        import org.springframework.web.bind.annotation.PostMapping;
-        import org.springframework.web.bind.annotation.RequestParam;
+        import org.springframework.web.bind.annotation.*;
 
+        import javax.persistence.criteria.CriteriaBuilder;
         import java.security.Principal;
         import java.util.ArrayList;
         import java.util.Optional;
@@ -21,6 +21,9 @@ public class MainController {
 
     @Autowired
     private PostsRepo postsRepo;
+
+    @Autowired
+    private  UsersRepo usersRepo;
 
     @Autowired
     private CustomUserDetailsService userDetailsService;
@@ -51,36 +54,77 @@ public class MainController {
     }
 
     @GetMapping("/main/{id}")
-    public String showPost(@PathVariable(value = "id") Integer id, Model model) {
+    public String showPost(@PathVariable(value = "id") Integer id, Model model, Principal principal) {
+
         Optional<Posts> post = postsRepo.findById(id);
+        Integer authorId = (post.get()).getAuthorId();
+        Integer currId = null;
+
+        if (principal != null) {
+
+            Users user = (Users) userDetailsService.loadUserByUsername(principal.getName());
+            currId = user.getId();
+        }
+
         ArrayList<Posts> res = new ArrayList<>();
         post.ifPresent(res::add);
+        model.addAttribute("authorId",authorId);
+        model.addAttribute("currId", currId);
         model.addAttribute("post",res);
         return "post-details";
     }
 
     @GetMapping("/main/{id}/edit")
-    public String editPost(@PathVariable(value = "id") Integer id, Model model) {
+    public String editPost(@PathVariable(value = "id") Integer id, Model model, Principal principal) {
+
         Optional<Posts> post = postsRepo.findById(id);
-        ArrayList<Posts> res = new ArrayList<>();
-        post.ifPresent(res::add);
-        model.addAttribute("post",res);
-        return "editPost";
+
+        Integer authorId = (post.get()).getAuthorId();
+        Integer currId = ((Users) userDetailsService.loadUserByUsername(principal.getName())).getId();
+
+        if ( authorId == currId) {
+
+            ArrayList<Posts> res = new ArrayList<>();
+            post.ifPresent(res::add);
+            model.addAttribute("post", res);
+            return "editPost";
+        }
+
+        return "redirect:/main";
     }
 
     @PostMapping("/main/{id}/edit")
-    public String updatePost(@PathVariable(value = "id") Integer id, @RequestParam String label, @RequestParam String postText, Model model) {
+    public String updatePost(@PathVariable(value = "id") Integer id, @RequestParam String label, @RequestParam String postText, Model model, Principal principal) {
+
         Posts post = postsRepo.findById(id).orElseThrow();
-        post.setLabel(label);
-        post.setPostText(postText);
-        postsRepo.save(post);
+        Integer authorId = post.getAuthorId();
+        Integer currId = ((Users) userDetailsService.loadUserByUsername(principal.getName())).getId();
+
+        if ( authorId == currId) {
+
+            post.setLabel(label);
+            post.setPostText(postText);
+            postsRepo.save(post);
+            return "redirect:/main";
+        }
+
         return "redirect:/main";
     }
 
     @PostMapping("/main/{id}/remove")
-    public String removePost(@PathVariable(value = "id") Integer id, Model model) {
+    public String removePost(@PathVariable(value = "id") Integer id, Model model, Principal principal) {
+
         Posts post = postsRepo.findById(id).orElseThrow();
-        postsRepo.delete(post);
+
+        Integer authorId = post.getAuthorId();
+        Integer currId = ((Users) userDetailsService.loadUserByUsername(principal.getName())).getId();
+
+        if ( authorId == currId) {
+
+            postsRepo.delete(post);
+            return "redirect:/main";
+        }
+
         return "redirect:/main";
     }
